Thaumcraft Aspect Registry
Added 2 New Aspects, Ascention and Terminus.

Added TC Aspects to all Avaritia Items that currently exist.

+++ src/main/java/morph/avaritia/client/render/shader/ShaderHelper.java
@@ -50,7 +50,7 @@ public final class ShaderHelper {
             int time = ARBShaderObjects.glGetUniformLocationARB(shader, "time");
 
             // Limits the animation time to a max value of 24000 ticks before resetting to 0
-            int animationTime = (int) (mc.player.ticksExisted % 24000);
+            int animationTime = mc.player.ticksExisted % 24000;
 
             if (mc.player != null && mc.player.world != null) {
